sys_fs_unlink: add CELL_EISDIR check

+++ rpcs3/Emu/Cell/lv2/sys_fs.cpp
@@ -852,6 +852,11 @@ error_code sys_fs_unlink(vm::cptr<char> path)
 		return {CELL_ENOTMOUNTED, path};
 	}
 
+	if (fs::is_dir(local_path))
+	{
+		return {CELL_EISDIR, path};
+	}
+
 	if (!vfs::host::unlink(local_path))
 	{
 		switch (auto error = fs::g_tls_error)
